Keep old message type (w) (#11)
To mark message as compressed we use the most significant bit of timestamp field.
When this bit is set, the timestamp just becomes negative (we will reset it on receiver), however the advantage is that there will be no problem with `pg_failover_slots` and order in `shared_preload_libraries`.

Besides that, solve a few minor problems.